requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
cchardet==2.1.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
Django==4.0.5
django-filter==21.1
django-simple-history==3.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
elastic-transport==8.1.2
elasticsearch==8.2.2
fake-useragent==0.1.11
frozenlist==1.3.0
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
idna==3.3
Markdown==3.3.7
multidict==6.0.2
numpy==1.22.4
oauthlib==3.2.0
pandas==1.4.2
pycares==4.1.2
pycparser==2.21
PyJWT==2.4.0
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==2.0.3
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tweepy==4.10.0
twint==2.1.20
urllib3==1.26.9
yarl==1.7.2
